Move api to own file
