+++ src/components/api/events/events.gateway.ts
@@ -10,7 +10,7 @@ import WebSocket from "ws";
 
 import { Event } from "./entities/event.entity";
 import { MQTT } from "../../mqtt";
-import { startServer, stopServer } from "../../main";
+import { startServer, stopServer } from "..";
 import { WsAuthGuard } from "../wsAuth.guard";
 import logger from "../../logger";
